+++ src/backend/InvenTree/company/models.py
@@ -650,6 +650,7 @@ class SupplierPartManager(models.Manager):
 
 
 class SupplierPart(
+    InvenTree.models.InvenTreeAttachmentMixin,
     InvenTree.models.MetadataMixin,
     InvenTree.models.InvenTreeBarcodeMixin,
     InvenTree.models.InvenTreeNotesMixin,