+++ src/frontend/src/pages/company/SupplierPartDetail.tsx
@@ -31,6 +31,7 @@ import {
 } from '../../components/items/ActionDropdown';
 import InstanceDetail from '../../components/nav/InstanceDetail';
 import { PageDetail } from '../../components/nav/PageDetail';
+import AttachmentPanel from '../../components/panels/AttachmentPanel';
 import NotesPanel from '../../components/panels/NotesPanel';
 import type { PanelType } from '../../components/panels/Panel';
 import { PanelGroup } from '../../components/panels/PanelGroup';
@@ -197,7 +198,7 @@ export default function SupplierPartDetail() {
         icon: 'purchase_orders'
       },
       {
-        type: 'string',
+        type: 'number',
         name: 'available',
         label: t`Supplier Availability`,
         hidden: !data.availability_updated,
@@ -205,7 +206,7 @@ export default function SupplierPartDetail() {
         icon: 'packages'
       },
       {
-        type: 'string',
+        type: 'date',
         name: 'availability_updated',
         label: t`Availability Updated`,
         copy: true,
@@ -277,6 +278,10 @@ export default function SupplierPartDetail() {
           <Skeleton />
         )
       },
+      AttachmentPanel({
+        model_type: ModelType.supplierpart,
+        model_id: supplierPart?.pk
+      }),
       NotesPanel({
         model_type: ModelType.supplierpart,
         model_id: supplierPart?.pk