+++ src/frontend/src/pages/part/PartDetail.tsx
@@ -47,7 +47,7 @@ import { UserRoles } from '@lib/enums/Roles';
 import { apiUrl } from '@lib/functions/Api';
 import { getDetailUrl } from '@lib/functions/Navigation';
 import { ActionButton } from '@lib/index';
-import type { ApiFormFieldSet, StockOperationProps } from '@lib/types/Forms';
+import type { StockOperationProps } from '@lib/types/Forms';
 import AdminButton from '../../components/buttons/AdminButton';
 import { PrintingActions } from '../../components/buttons/PrintingActions';
 import StarredToggleButton from '../../components/buttons/StarredToggleButton';
@@ -184,7 +184,7 @@ function BomValidationInformation({
         <Text>{t`Do you want to validate the bill of materials for this assembly?`}</Text>
       </Alert>
     ),
-    successMessage: t`BOM validated`,
+    successMessage: t`Bill of materials scheduled for validation`,
     onFormSuccess: () => {
       bomInformationQuery.refetch();
     }
@@ -1051,38 +1051,10 @@ export default function PartDetail() {
     onFormSuccess: refreshInstance
   });
 
-  const createPartFields = usePartFields({ create: true });
-
-  const duplicatePartFields: ApiFormFieldSet = useMemo(() => {
-    return {
-      ...createPartFields,
-      duplicate: {
-        children: {
-          part: {
-            value: part.pk,
-            hidden: true
-          },
-          copy_images: {
-            value: true
-          },
-          copy_bom: {
-            value: part.assembly && globalSettings.isSet('PART_COPY_BOM'),
-            hidden: !part.assembly
-          },
-          copy_notes: {
-            value: true
-          },
-          copy_parameters: {
-            value: globalSettings.isSet('PART_COPY_PARAMETERS')
-          },
-          copy_tests: {
-            value: part.testable,
-            hidden: !part.testable
-          }
-        }
-      }
-    };
-  }, [createPartFields, globalSettings, part]);
+  const duplicatePartFields = usePartFields({
+    create: true,
+    duplicatePartInstance: part
+  });
 
   const duplicatePart = useCreateApiFormModal({
     url: ApiEndpoints.part_list,